requirements.txt
discord.py>=2.4,<3
asyncpg==0.30.0
gspread==6.2.1
google-auth>=2.35
python-dotenv==1.0.1
aiohttp>=3.10
